clean up junk from mela.py into a separate file
